11233 Removing innerText & adding eslint rule
