+++ plugins/eslint/package.json
@@ -2,7 +2,7 @@
   "name": "eslint-plugin-prebid",
   "version": "1.0.0",
   "description": "validates module imports can be found without custom webpack resolvers, are in module whitelist, and not module entry points",
-  "main": "validateImports.js",
+  "main": "index.js",
   "author": "the prebid.js contributors",
   "license": "Apache-2.0"
 }
